do not try to clean a non existing destination directory

+++ classes/report/fields/runner/coverage/html.php
@@ -82,7 +82,7 @@ class html extends report\fields\runner\coverage\cli
 
 	public function __toString()
 	{
-        //try {
+        try {
             $string = parent::__toString();
 
             if ($this->adapter->extension_loaded('xdebug') === true)
@@ -317,9 +317,9 @@ class html extends report\fields\runner\coverage\cli
 
                 $string .= $this->urlPrompt . $this->urlColorizer->colorize(sprintf($this->locale->_('Details of code coverage are available at %s.'), $this->rootUrl)) . PHP_EOL;
             }
-//        } catch (\Exception $e) {
-//            trigger_error($e->getMessage(), E_USER_ERROR);
-//        }
+        } catch (\Exception $e) {
+            trigger_error($e->getMessage(), E_USER_ERROR);
+        }
 
 		return $string;
 	}
@@ -465,17 +465,20 @@ class html extends report\fields\runner\coverage\cli
 
 	public function cleanDestinationDirectory()
 	{
-		foreach ($this->getDestinationDirectoryIterator() as $inode)
-		{
-			if ($inode->isDir() === false)
-			{
-				$this->adapter->unlink($inode->getPathname());
-			}
-			else if (($pathname = $inode->getPathname()) !== $this->destinationDirectory)
-			{
-				$this->adapter->rmdir($pathname);
-			}
-		}
+        if ($this->adapter->is_dir($this->destinationDirectory))
+        {
+            foreach ($this->getDestinationDirectoryIterator() as $inode)
+            {
+                if ($inode->isDir() === false)
+                {
+                    $this->adapter->unlink($inode->getPathname());
+                }
+                else if (($pathname = $inode->getPathname()) !== $this->destinationDirectory)
+                {
+                    $this->adapter->rmdir($pathname);
+                }
+            }
+        }
 
 		return $this;
 	}
